feat: add report feature and backend adjustments

+++ backend/wall-pojo/src/main/java/cn/yiming1234/NottinghamWall/dto/AdminDTO.java
@@ -7,16 +7,11 @@ import java.io.Serializable;
 @Data
 public class AdminDTO implements Serializable {
 
-    private Long id;
-
+    private Integer id;
     private String username;
-
     private String name;
-
     private String phone;
-
     private String sex;
-
     private String idNumber;
 
 }
